Move Deciders interface for consistency.
