+++ pkg/reconciler/autoscaling/kpa/kpa.go
@@ -49,24 +49,6 @@ import (
 	"k8s.io/client-go/tools/cache"
 )
 
-// Deciders is an interface for notifying the presence or absence of KPAs.
-type Deciders interface {
-	// Get accesses the Decider resource for this key, returning any errors.
-	Get(ctx context.Context, namespace, name string) (*autoscaler.Decider, error)
-
-	// Create adds a Decider resource for a given key, returning any errors.
-	Create(ctx context.Context, decider *autoscaler.Decider) (*autoscaler.Decider, error)
-
-	// Delete removes the Decider resource for a given key, returning any errors.
-	Delete(ctx context.Context, namespace, name string) error
-
-	// Watch registers a function to call when Decider change.
-	Watch(watcher func(string))
-
-	// Update update the Decider resource, return the new Decider or any errors.
-	Update(ctx context.Context, decider *autoscaler.Decider) (*autoscaler.Decider, error)
-}
-
 // Reconciler tracks PAs and right sizes the ScaleTargetRef based on the
 // information from Deciders.
 type Reconciler struct {
@@ -75,7 +57,7 @@ type Reconciler struct {
 	serviceLister   corev1listers.ServiceLister
 	sksLister       nlisters.ServerlessServiceLister
 	endpointsLister corev1listers.EndpointsLister
-	kpaDeciders     Deciders
+	kpaDeciders     resources.Deciders
 	metrics         aresources.Metrics
 	scaler          *scaler
 	configStore     configStore
